Move TOML configuration related definitions in a sub-module
